kbuild: Implement the inline optimizations configuration
Specify increased inlining to provide an improvement in performance. Though it will lead to an increase in the size of the final kernel image, this configuration provides a bit more control over how much is inlined to further optimise the kernel for certain workloads.

Note: Despite using selective inline optimisation, this is much lower and worth using generally for any missed sensitive areas.

Signed-off-by: prathamdubey2005 <134331217+prathamdubey2005@users.noreply.github.com>